+++ src/cose/receipt/consistency/index.ts
