--- src/spice/index.ts
@@ -1,3 +1,0 @@
-import * as receipt from './receipt'
-
-export { receipt }
